+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to [EffCSS](https://github.com/msabitov/effcss) are document
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.4.0] - 2026-07-17
+
+### Added
+
+- `subscribe` utility that tracks the EffCSS style creation events (types and tests added)
+
+### Changed
+
+- minor adjustments to coverage settings
+
 ## [5.3.1] - 2026-07-10
 
 ### Added
@@ -723,7 +733,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-[5.3.1]: https://github.com/msabitov/effcss
+[5.4.0]: https://github.com/msabitov/effcss
+[5.3.1]: https://github.com/msabitov/effcss/commit/f4d3b36045c1d2f11337c8f5489ce8521ba41dec
 [5.3.0]: https://github.com/msabitov/effcss/commit/ea8e29f4a2cfe3d0b9e2a82a49d7c9317df25c5f
 [5.2.0]: https://github.com/msabitov/effcss/commit/583f58540b402507a3573de8658cd7b3d6529415
 [5.1.0]: https://github.com/msabitov/effcss/commit/7f5f831225fe8b4b5ba5f67c9568f3bd798a1ce8